+++ extensions/broker/src/main/java/de/sovity/extension/broker/sender/UnregisterResourceRequestSender.java
@@ -1,0 +1,69 @@
+/*
+ *  Copyright (c) 2022 sovity GmbH
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       sovity GmbH - initial API and implementation
+ *
+ */
+package de.sovity.extension.broker.sender;
+
+import de.fraunhofer.iais.eis.DynamicAttributeToken;
+import de.fraunhofer.iais.eis.Message;
+import de.fraunhofer.iais.eis.MessageProcessedNotificationMessageImpl;
+import de.fraunhofer.iais.eis.ResourceUnavailableMessageBuilder;
+import de.sovity.extension.broker.sender.message.UnregisterResourceMessage;
+import org.eclipse.edc.protocol.ids.api.multipart.dispatcher.sender.MultipartSenderDelegate;
+import org.eclipse.edc.protocol.ids.api.multipart.dispatcher.sender.response.IdsMultipartParts;
+import org.eclipse.edc.protocol.ids.api.multipart.dispatcher.sender.response.MultipartResponse;
+import org.eclipse.edc.protocol.ids.spi.domain.IdsConstants;
+import org.eclipse.edc.protocol.ids.util.CalendarUtil;
+
+import java.util.List;
+
+import static org.eclipse.edc.protocol.ids.api.multipart.dispatcher.sender.util.ResponseUtil.parseMultipartStringResponse;
+import static org.eclipse.edc.protocol.ids.jsonld.JsonLd.getObjectMapper;
+
+public class UnregisterResourceRequestSender implements MultipartSenderDelegate<UnregisterResourceMessage, String> {
+
+    public UnregisterResourceRequestSender() {
+    }
+
+    @Override
+    public Message buildMessageHeader(UnregisterResourceMessage unregisterResourceMessage,
+                                      DynamicAttributeToken token) throws Exception {
+        return new ResourceUnavailableMessageBuilder()
+                ._modelVersion_(IdsConstants.INFORMATION_MODEL_VERSION)
+                ._issued_(CalendarUtil.gregorianNow())
+                ._securityToken_(token)
+                ._issuerConnector_(unregisterResourceMessage.connectorBaseUrl())
+                ._senderAgent_(unregisterResourceMessage.connectorBaseUrl())
+                ._affectedResource_(unregisterResourceMessage.affectedResourceUri())
+                .build();
+    }
+
+    @Override
+    public String buildMessagePayload(UnregisterResourceMessage unregisterResourceMessage) throws Exception {
+        return null;
+    }
+
+    @Override
+    public MultipartResponse<String> getResponseContent(IdsMultipartParts parts) throws Exception {
+        return parseMultipartStringResponse(parts, getObjectMapper());
+    }
+
+    @Override
+    public List<Class<? extends Message>> getAllowedResponseTypes() {
+        return List.of(MessageProcessedNotificationMessageImpl.class);
+    }
+
+    @Override
+    public Class<UnregisterResourceMessage> getMessageType() {
+        return UnregisterResourceMessage.class;
+    }
+}
