+++ extensions/broker/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-BrokerExtension
+de.sovity.extension.broker.BrokerExtension
