+++ extensions/clearinghouse/src/main/java/de/sovity/extension/clearinghouse/sender/message/LogMessage.java
@@ -1,0 +1,22 @@
+package de.sovity.extension.clearinghouse.sender.message;
+
+import de.sovity.extension.clearinghouse.sender.message.clearingdispatcher.ExtendedMessageProtocolClearing;
+import org.eclipse.edc.connector.contract.spi.types.agreement.ContractAgreement;
+import org.eclipse.edc.spi.types.domain.message.RemoteMessage;
+
+import java.net.URI;
+import java.net.URL;
+
+public record LogMessage(URL clearingHouseLogUrl,
+                         URI connectorBaseUrl,
+                         ContractAgreement contractAgreement) implements RemoteMessage {
+    @Override
+    public String getProtocol() {
+        return ExtendedMessageProtocolClearing.IDS_EXTENDED_PROTOCOL_CLEARING;
+    }
+
+    @Override
+    public String getConnectorAddress() {
+        return clearingHouseLogUrl.toString();
+    }
+}
