+++ extensions/clearinghouse/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-ClearingHouseExtension
+de.sovity.extension.clearinghouse.ClearingHouseExtension
